refactor(chat): extract the channel family to @endo/space-channel
The final bucket of the chat/channel split. Move the four multiuser
channel-view bodies (channel, forum, microblog, outliner) plus their
shared substrate (channel-utils, react-utils, edit-queue, profile-popup)
into the new @endo/space-channel package, as a unit — so the substrate
coupling that pinned them resolves: a body imports channel-utils from
within its own package, not back into the @endo/chat app.

Boundary decisions, per the split contract:
- channel-header stays in the shell (it needs heat-*, not the moved
  substrate, and is dispatch-root chrome, not a body).
- send-form stays in the shell; the bodies referenced only its
  SendFormAPI *type* (an opaque chatBarAPI handle), now loosened to
  `object` so space-channel carries no send-form dependency.
- token-autocomplete already moved to @endo/chat-kit (Phase A).
- channel.css ships via the package's ./channel.css export (imported in
  main.js, like peers/floot), replacing the index.html <link>.

chat.js imports the four bodies and the ChannelMessage type from
@endo/space-channel; share-modal repoints its type import. The bodies
keep their entry signatures, so the dispatch root is otherwise untouched.
Build, both packages' eslint, and the 106 channel-family tests pass.
