+++ packages/chat/test/component/channel.test.js
@@ -6,8 +6,8 @@ import '@endo/init/debug.js';
 import test from 'ava';
 import { Far } from '@endo/far';
 import { readerFromIterator } from '@endo/exo-stream/reader-from-iterator.js';
+import { channelComponent } from '@endo/space-channel/channel-component.js';
 import { createDOM, tick } from '../helpers/dom-setup.js';
-import { channelComponent } from '../../channel-component.js';
 
 // The default multiuser channel body view, migrated from imperative DOM to a
 // confined Preact component rendered through `renderConfined`. It composes the
