+++ packages/chat/test/component/microblog.test.js
@@ -6,8 +6,8 @@ import '@endo/init/debug.js';
 import test from 'ava';
 import { Far } from '@endo/far';
 import { readerFromIterator } from '@endo/exo-stream/reader-from-iterator.js';
+import { microblogComponent } from '@endo/space-channel/microblog-component.js';
 import { createDOM, tick } from '../helpers/dom-setup.js';
-import { microblogComponent } from '../../microblog-component.js';
 
 // The microblog body view, migrated from imperative DOM to a confined Preact
 // component rendered through `renderConfined`. It composes the reused imperative
