+++ packages/space-channel/package.json
@@ -1,0 +1,63 @@
+{
+  "name": "@endo/space-channel",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Multiuser channel space family for Endo Chat — the channel, forum, outliner, and microblog views plus their shared channel/react/edit substrate",
+  "keywords": [
+    "endo",
+    "chat",
+    "channel",
+    "preact"
+  ],
+  "author": "Endo contributors",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/endojs/endo.git",
+    "directory": "packages/space-channel"
+  },
+  "bugs": {
+    "url": "https://github.com/endojs/endo/issues"
+  },
+  "type": "module",
+  "exports": {
+    ".": "./index.js",
+    "./channel-component.js": "./src/channel-component.js",
+    "./channel-utils.js": "./src/channel-utils.js",
+    "./channel.css": "./src/channel.css",
+    "./edit-queue.js": "./src/edit-queue.js",
+    "./forum-component.js": "./src/forum-component.js",
+    "./microblog-component.js": "./src/microblog-component.js",
+    "./outliner-component.js": "./src/outliner-component.js",
+    "./package.json": "./package.json",
+    "./profile-popup.js": "./src/profile-popup.js",
+    "./react-utils.js": "./src/react-utils.js"
+  },
+  "scripts": {
+    "build": "exit 0",
+    "lint": "yarn lint:eslint",
+    "lint-fix": "yarn lint:eslint --fix",
+    "lint:eslint": "eslint '**/*.js'"
+  },
+  "dependencies": {
+    "@endo/chat-kit": "workspace:^",
+    "@endo/daemon": "workspace:^",
+    "@endo/exo-stream": "workspace:^",
+    "@endo/far": "workspace:^",
+    "@endo/harden": "workspace:^",
+    "@endo/monaco-wrapper": "workspace:^",
+    "@endo/preact-container": "workspace:^",
+    "preact": "^10.29.2"
+  },
+  "devDependencies": {
+    "eslint": "catalog:dev"
+  },
+  "eslintConfig": {
+    "extends": [
+      "plugin:@endo/internal"
+    ],
+    "env": {
+      "browser": true
+    }
+  }
+}
