+++ packages/space-channel/src/channel-component.js
