+++ packages/space-channel/src/forum-component.js
