+++ packages/space-channel/src/profile-popup.js
